requirements.txt
pycryptodome
sqlcipher3-wheels
python-magic
python-magic-bin